updates complete up to experiment 6
