+++ examples/sparkfun_redbot/sparkfun_experiments/Exp2_DriveForward.py
@@ -11,8 +11,6 @@
 
 from pymata_aio.pymata3 import PyMata3
 from library.redbot import RedBotMotors
-import sys
-import signal
 # This line "includes" the RedBot library into your sketch.
 # Provides special objects, methods, and functions for the RedBot.
 
@@ -29,17 +27,7 @@ motors = RedBotMotors(board)
 # Instantiate the motor control object. This only needs to be done once.
 
 
-def signal_handler(sig, frame):
-    """Helper method to shutdown the RedBot if Ctrl-c is pressed"""
-    print('\nYou pressed Ctrl+C')
-    if board is not None:
-        board.send_reset()
-        board.shutdown()
-    sys.exit(0)
-
-
 def setup():
-    signal.signal(signal.SIGINT, signal_handler)
     print("Left and right motors at full speed forward")
     motors.drive(255)   # Turn on Left and right motors at full speed forward.
     board.sleep(2.0)    # Waits for 2 seconds
