+++ examples/sparkfun_redbot/sparkfun_experiments/Exp3_Turning.py
@@ -10,10 +10,11 @@
   connected, and the board must be receiving power from the battery pack.
  """
 
-from pymata_aio.pymata3 import PyMata3
-from library.redbot import RedBotMotors
 import sys
 import signal
+
+from pymata_aio.pymata3 import PyMata3
+from library.redbot import RedBotMotors
 # This line "includes" the RedBot library into your sketch.
 # Provides special objects, methods, and functions for the RedBot.
