+++ examples/sparkfun_redbot/sparkfun_experiments/Exp5_Bumpers.py
@@ -14,17 +14,25 @@
 
   8 Oct 2013 M. Hord
   Revised 30 Oct 2014 B. Huang
+  Revised 2 Oct 2015 L. Mathews
 """
+
+import sys
+import signal
 
 from pymata_aio.pymata3 import PyMata3
 from library.redbot import RedBotMotors, RedBotBumper
-from pymata_aio.constants import Constants
-
-COM_PORT = None # Use automatic com port detection (the default)
-#COM_PORT = "COM7" # Manually specify the com port (optional)
 
 
-board = PyMata3(com_port=COM_PORT)
+WIFLY_IP_ADDRESS = None            # Leave set as None if not using WiFly
+WIFLY_IP_ADDRESS = "137.112.217.88"  # If using a WiFly on the RedBot, set the ip address here.
+if WIFLY_IP_ADDRESS:
+  board = PyMata3(ip_address=WIFLY_IP_ADDRESS)
+else:
+  # Use a USB cable to RedBot or an XBee connection instead of WiFly.
+  COM_PORT = None # Use None for automatic com port detection, or set if needed i.e. "COM7"
+  board = PyMata3(com_port=COM_PORT)
+
 # Instantiate the motor control object. This only needs to be done once.
 motors = RedBotMotors(board)
 
@@ -34,23 +42,33 @@ right_bumper = RedBotBumper(board, 11)  # initializes bumper object on pin 11
 BUTTON_PIN = 12
 
 
+def signal_handler(sig, frame):
+    """Helper method to shutdown the RedBot if Ctrl-c is pressed"""
+    print('\nYou pressed Ctrl+C')
+    if board is not None:
+        board.send_reset()
+        board.shutdown()
+    sys.exit(0)
+
+
 def setup():
-    # nothing here
-    pass
+    signal.signal(signal.SIGINT, signal_handler)
+    print("Experiment 5 - Bump sensors")
 
 
 def loop():
     motors.drive(255)
+    board.sleep(0.1)  # When using a wireless connection a small sleep is necessary
 
     left_bumper_state = left_bumper.read()
-    # board.sleep(0.1)  # When using XBee a small sleep is necessary
     right_bumper_state = right_bumper.read()
-    # board.sleep(0.1)  # When using XBee a small sleep is necessary
     if left_bumper_state == 0: # left bumper is bumped
+        print("Left bump")
         reverse()
         turn_right()
 
     if right_bumper_state == 0: # left bumper is bumped
+        print("Right bump")
         reverse()
         turn_left()
